Move remark parser commands to new package
