+++ src/main/java/io/github/mivek/parser/command/remark/SmallHailSizeCommand.java
@@ -1,4 +1,4 @@
-package io.github.mivek.parser.remark;
+package io.github.mivek.parser.command.remark;
 
 import io.github.mivek.internationalization.Messages;
 import io.github.mivek.parser.RemarkParser;
@@ -7,20 +7,21 @@ import io.github.mivek.utils.Regex;
 /**
  * @author mivek
  */
-public class HailSizeCommand implements Command {
+public class SmallHailSizeCommand implements Command {
+
     /** The message instance. */
     private final Messages fMessages;
 
     /**
      * Default constructor.
      */
-    public HailSizeCommand() {
+    public SmallHailSizeCommand() {
         fMessages = Messages.getInstance();
     }
 
     @Override public final String execute(final String pRemark, final StringBuilder pStringBuilder) {
-        String[] hailParts = Regex.pregMatch(RemarkParser.HAIL_SIZE, pRemark);
-        pStringBuilder.append(fMessages.getString("Remark.Hail", hailParts[1])).append(" ");
-        return pRemark.replaceFirst(RemarkParser.HAIL_SIZE.pattern(), "").trim();
+        String[] hailParts = Regex.pregMatch(RemarkParser.HAIL_SIZE_LESS_THAN, pRemark);
+        pStringBuilder.append(fMessages.getString("Remark.Hail.LesserThan", hailParts[1])).append(" ");
+        return pRemark.replaceFirst(RemarkParser.HAIL_SIZE_LESS_THAN.pattern(), "").trim();
     }
 }
