+++ src/main/java/io/github/mivek/parser/command/remark/WindShiftFropaCommand.java
@@ -1,4 +1,4 @@
-package io.github.mivek.parser.remark;
+package io.github.mivek.parser.command.remark;
 
 import io.github.mivek.internationalization.Messages;
 import io.github.mivek.parser.RemarkParser;
@@ -7,7 +7,7 @@ import io.github.mivek.utils.Regex;
 /**
  * @author mivek
  */
-public class WindPeakCommand implements Command {
+public class WindShiftFropaCommand implements Command {
 
     /** The messages instance. */
     private final Messages fMessages;
@@ -15,14 +15,14 @@ public class WindPeakCommand implements Command {
     /**
      * Default constructor.
      */
-    public WindPeakCommand() {
+    public WindShiftFropaCommand() {
         fMessages = Messages.getInstance();
     }
 
     @Override public final String execute(final String pRemark, final StringBuilder pStringBuilder) {
-        String[] windPeakParts = Regex.pregMatch(RemarkParser.WIND_PEAK, pRemark);
-        pStringBuilder.append(fMessages.getString("Remark.PeakWind", windPeakParts[1], windPeakParts[2], verifyString(windPeakParts[3]), windPeakParts[4]));
+        String[] windShiftParts = Regex.pregMatch(RemarkParser.WIND_SHIFT_FROPA, pRemark);
+        pStringBuilder.append(fMessages.getString("Remark.WindShift.FROPA", verifyString(windShiftParts[1]), windShiftParts[2]));
         pStringBuilder.append(" ");
-        return pRemark.replaceFirst(RemarkParser.WIND_PEAK.pattern(), "").trim();
+        return pRemark.replaceFirst(RemarkParser.WIND_SHIFT_FROPA.pattern(), "");
     }
 }
