+++ src/main/java/io/github/mivek/parser/command/remark/package-info.java
@@ -2,4 +2,4 @@
  * @author mivek
  * This package contains command related to remark parser.
  */
-package io.github.mivek.parser.remark;
+package io.github.mivek.parser.command.remark;
